module: drop support for extensionless main entry points in esm
PR-URL: https://github.com/nodejs/node/pull/31415
