+++ lib/internal/modules/esm/get_format.js
@@ -54,11 +54,13 @@ function defaultGetFormat(url, context, defaultGetFormat) {
     return { format };
   } else if (parsed.protocol === 'file:') {
     const ext = extname(parsed.pathname);
-    let format = extensionFormatMap[ext];
-    const isMain = context.parentURL === undefined;
-    if (ext === '.js' || (!format && isMain))
+    let format;
+    if (ext === '.js') {
       format = getPackageType(parsed.href) === TYPE_MODULE ?
         'module' : 'commonjs';
+    } else {
+      format = extensionFormatMap[ext];
+    }
     if (!format) {
       if (experimentalSpeciferResolution === 'node') {
         process.emitWarning(
