+++ lib/internal/modules/esm/loader.js
@@ -96,7 +96,7 @@ class Loader {
     }
 
     const getFormatResponse = await this._getFormat(
-      url, { parentURL }, defaultGetFormat);
+      url, {}, defaultGetFormat);
     if (typeof getFormatResponse !== 'object') {
       throw new ERR_INVALID_RETURN_VALUE(
         'object', 'loader getFormat', getFormatResponse);
