+++ test/fixtures/es-modules/package-type-module/imports-noext.mjs
@@ -1,0 +1,1 @@
+import './noext-esm';
