+++ test/fixtures/es-modules/package-type-module/imports-unknownext.mjs
@@ -1,0 +1,1 @@
+import './extension.unknown';
